Client (including dist) and server folders developed
Developed the client folder in order to create a dist folder when starting the code. This will generate multiple bundle files (database, editor, header, install, main and a manifest). This allows to be seen when you inspect the console on the server. Further, the server with the scripts to run concurrently both the server and clinet,